Add files via upload
This is my first time attempting GitHub so I'm not sure if I put up the right file. This is what I have so far. This programs asks for some details and then generates a math worksheet. In future iterations I would like to conceal the answers and reveal them with a button click after generation. Also I couldn't figure out how to link this to an html template in PyCharm.